requirements.txt
# Core dependencies
torch==1.12.1
torchaudio==0.12.1
transformers==4.21.3
requests==2.27.1
numpy==1.22.4
soundfile==0.10.3
librosa==0.9.2
vosk==0.3.43
datasets==2.4.0

pydantic==1.9.2
langchain==0.0.240

# ROS2 dependencies
rclpy
geometry_msgs
std_msgs

# Audio processing
sounddevice==0.4.5

# Development environment (optional)
IPython==8.4.0
jupyter==1.0.0

# Environment management
python-dotenv==0.19.0